Sticking to OOP and SOLID [Shopping Items]
Replace abstract class with interfaces
