--- src/Demo/CartBundle/Entity/AbstractShoppingItem.php
@@ -1,128 +1,0 @@
-<?php
-
-
-namespace Demo\CartBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * Class AbstractShoppingItem
- * @package Demo\CartBundle\Entity
- *
- * Base class for Shopping items
- *
- * @ORM\MappedSuperclass
- */
-abstract class AbstractShoppingItem
-{
-    /**
-     * @ORM\Column(type="integer")
-     * @ORM\Id
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @ORM\Column(type="string", length=255)
-     */
-    protected $name;
-
-    /**
-     * @ORM\Column(type="text")
-     */
-    protected $description;
-
-    /**
-     * @ORM\Column(type="decimal", scale=2)
-     */
-    protected $price;
-
-    /**
-     * @ORM\Column(type="boolean")
-     */
-    protected $inStock;
-
-    /**
-     * @return mixed
-     */
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getName()
-    {
-        return $this->name;
-    }
-
-    /**
-     * @param mixed $name
-     */
-    public function setName($name)
-    {
-        $this->name = $name;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getDescription()
-    {
-        return $this->description;
-    }
-
-    /**
-     * @param mixed $description
-     */
-    public function setDescription($description)
-    {
-        $this->description = $description;
-    }
-
-    /**
-     * returns the actual price of the item
-     *
-     * @return mixed
-     */
-    public function getPrice()
-    {
-        return $this->price;
-    }
-
-    /**
-     * @param mixed $price
-     */
-    public function setPrice($price)
-    {
-        $this->price = $price;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getInStock()
-    {
-        return $this->inStock;
-    }
-
-    /**
-     * @param mixed $inStock
-     */
-    public function setInStock($inStock)
-    {
-        $this->inStock = $inStock;
-    }
-
-    /**
-     * returns the net price of an item after discount(if exists) or other item price manipulation
-     *
-     * @return mixed
-     */
-    public function getNetPrice()
-    {
-        return $this->getPrice();
-    }
-}
