+++ src/Demo/CartBundle/Entity/OrderElement/OrderItem.php
@@ -2,7 +2,7 @@
 
 namespace Demo\CartBundle\Entity\OrderElement;
 
-use Demo\CartBundle\Entity\AbstractShoppingItem;
+use Demo\CartBundle\Entity\ShoppingItemInterface;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
@@ -45,7 +45,7 @@ class OrderItem implements OrderItemInterface
     }
 
     /**
-     * @return AbstractShoppingItem
+     * @return ShoppingItemInterface
      */
     public function getItem()
     {
@@ -55,7 +55,7 @@ class OrderItem implements OrderItemInterface
     /**
      * @param mixed $item
      */
-    public function setItem(AbstractShoppingItem $item)
+    public function setItem(ShoppingItemInterface $item)
     {
         $this->item = $item;
     }
